Kumar Sir Class 12 Nov : Binary Search Tree Continued

+++ kumar-gaurav-sir/2024-11-05/BST.java
@@ -92,7 +92,7 @@ public class BST
 			{
 				if (word.length() < 10) continue;
 
-				Integer curr_count = binarySearchTreeWordCount.get(word);
+				Integer curr_count = (Integer) binarySearchTreeWordCount.get(word);
 
 				if (curr_count != null)
 				{
